requirements.txt
kafka-python
pytest 